+++ console-api/cmd/server/iam_policies.go
@@ -34,7 +34,7 @@ var policyResources = []string{
 	"fileshares", "directories", "migrations", "replications", "dataflows", "streams",
 	"securitygroups", "faultinjections", "queries", "httpapis", "graphqlapis",
 	"databaseproxies", "statemachines", "trainingjobs", "modelpackages", "batchtransforms", "processingjobs", "modelmonitors", "featuregroups",
-	"staticsites", "parameters",
+	"staticsites", "parameters", "emailsenders",
 }
 
 // policyVerbs are the verbs an action may use (case-insensitive), plus "*".